+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- Metadata Retriever & Package XML viewer : Add clickable metadata-type links based on https://sf-explorer.github.io/sf-doc-to-json.
+- Metadata Retriever & Package XML viewer enhancements:
+  - Add clickable metadata-type links based on https://sf-explorer.github.io/sf-doc-to-json
+  - Add clickable icon to display standard objects documentation
 
 ## [6.22.0] 2025-12-26